pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-search-console-cli"
version = "0.1.0"
description = "CLI for Google Search Console"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8.1.7",
  "google-api-python-client>=2.140.0",
  "google-auth>=2.32.0",
  "google-auth-oauthlib>=1.2.1",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
gsc = "gsc_cli.cli:cli"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools]
packages = ["gsc_cli"]
